Moved UserKeys.kt into firebase package
